refactor(settings): collapse navigation state behind model.nav
SettingsModel.swift sat at 348/350 and @Published cannot live in
an extension, so the file hard-blocked any change touching model
state — and #277 part 2 needs ~10 more navigation fields (drawer
expansions, the Shortcuts mode strip).

Collapse the six transient navigation members plus their four
mutators into one value type, SettingsNavigation, behind a single
@Published var nav. Call sites keep their shape: $model.nav.barEditor
still projects a Binding, model.nav.startFlash(...) still mutates
and publishes.

Moved verbatim (behaviour-free): pendingReveal, pendingScroll,
flash (private(set)), flashToken, layoutModeTab, barEditor, and
resetSurfaces / resetLayoutModeTab / startFlash / endFlash. The
onChange(of:pendingReveal) placement on the outer Group (the sole
reason a Lua-mode request resolves later) and the one-writer /
one-clearer split across pendingReveal→pendingScroll are unchanged.
Model drops 348→267 lines.

The old SettingsNavigation.swift held only cross-reference-link
machinery (no type of that name), so it is renamed to
CrossReferenceLink.swift to free the type name for the new value
type.

Verify: swift build, 2088+14 tests, lint exit 0, release build all
green locally (CI Actions dead on billing).
